Fortified test case for sumproduct algorithm.
git-svn-id: https://svn.cc.gatech.edu/borg/OccupancyGrid/trunk@17860 898a188c-9671-0410-8e00-e3fd810bbb7f

+++ OccupancyGrid/include/visualiser.h
@@ -6,6 +6,7 @@
 #pragma once
 #include <opencv2/opencv.hpp>
 #include <boost/format.hpp>
+#include "LaserFactor.h"
 
 static const cv::Vec3f GREEN(0, 255, 0);
 static const cv::Vec3f WHITE(255, 255, 255);
@@ -39,15 +40,15 @@ public:
 
   /// Set Occupancy for visualization
   inline void setOccupancy(const LaserFactor::Occupancy& occ) {
-    for (size_t i = 0; i < vis_.rows; i++)
-      for (size_t j = 0; j < vis_.cols; j++)
+    for (int i = 0; i < vis_.rows; i++)
+      for (int j = 0; j < vis_.cols; j++)
         vis_.at<cv::Vec3b>(i, j) = (occ.at(i*vis_.cols + j) != 0) ? BLACK : WHITE;
   }
 
   inline void setMarginals(const std::vector<double>& marg) {
     double max_ele = *std::max_element(marg.begin(), marg.end());
-    for (size_t i = 0; i < vis_.rows; i++) {
-      for (size_t j = 0; j < vis_.cols; j++) {
+    for (int i = 0; i < vis_.rows; i++) {
+      for (int j = 0; j < vis_.cols; j++) {
         uint8_t val = (max_ele - marg.at(i*vis_.cols + j)) * 255 / max_ele;
         vis_.at<cv::Vec3b>(i, j) = cv::Vec3b(val, val, val);
       }
@@ -56,7 +57,7 @@ public:
 
   /// Highlight given vector of cells
   inline void highlightCells(const std::vector<gtsam::Index> & cells) {
-    for (int i = 0; i < cells.size(); i++) {
+    for (size_t i = 0; i < cells.size(); i++) {
       gtsam::Index idx = cells[i];
       size_t x,y;
       idx2xy(idx, x, y);
